refactor: import utils functions from utils.rs and make them available from anywhere
